Added more libraries to security layer

+++ create_env_file.sh
@@ -1,3 +1,14 @@
+#!/bin/bash
+
+# .env file
+filename=.env
+
 # Create a new file with .env initial values
-cp .env.test .env
-echo ".env file created"
+if [ ! -f $fileame ]; 
+then
+    echo ".env File not found, creating new one..."
+    cp .env.test .env
+    echo ".env File created!"
+fi
+
+
